+++ oauth2_provider/authorization_server/views/__init__.py
@@ -1,0 +1,2 @@
+"""Authorization Server view layer (authorization/token/revocation/device/DCR
+endpoints, server metadata, and management UI)."""
