+++ oauth2_provider/resource_server/views/__init__.py
@@ -1,0 +1,1 @@
+"""Resource Server view layer (protected-resource views and RFC 9728 metadata)."""
